Cambios a DAO de articulos

+++ src/integracion/DAOCategorias.java
@@ -3,9 +3,13 @@ package integracion;
 import java.util.List;
 
 public interface DAOCategorias {
-    public void altaArticuloCat(int id, String fechal, int descuento, String cat);
+    public void altaArticuloCat(int id, String fechal, double descuento, String cat);
+
     public void bajaArticuloCat(int id);
+
     public void modificarArticulo(int id, String fechal, int descuento, String cat);
+
     public List<String> getCategorias();
+
     public void actualizaExclusivos();
 }
